server/requirements.txt
flask 
flask-cors 
torch 
easyocr 
opencv-python-headless 
numpy 
pillow 
requests 
pymupdf 
torchvision
python-dotenv
flask-bcrypt 
flask-jwt-extended 
pymongo
twilio
